Simplify encryption error predicates for signed-in non-syncing users
No behavior change.

Before crrev.com/c/4874142, the opt-in for passwords-account-storage
on desktop was implemented via
DataTypeController::GetPreconditionState(), rather than SyncUserSettings::GetSelectedTypes(). As a consequence, for users
who hadn't opted in, PASSWORDS was nevertheless considered
"preferred", and the following could return true:
1. SyncUserSettings::IsTrustedVaultKeyRequiredForPreferredDataTypes()
2. SyncUserSettings::IsTrustedVaultRecoverabilityDegraded()
This required special helpers:
3. ShouldShowSyncKeysMissingError()
4. ShouldShowTrustedVaultDegradedRecoverabilityError()
which checked for the opt-in on top of 1/2.

Now 1 and 2 just work, this CL removes 3 and 4.

Fixed: 1447083
Change-Id: I488e65a63a7d4d6ea34a0c510cf77ff2ead30b13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5049884
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1229311}