Make periodic logs synchronous when shutting down
Full context:
Currently, we finalize periodic logs asynchronously. This includes
things like serializing the log. Once that's done, a "reply task" is
posted back on the main thread to store that finalized log in Local
State.

One possible issue is that while the background task to finalize the
log is running, the user may shut down the browser. When this happens,
the "reply task" is never run, so the finalized log is lost. For this
reason, we don't do "delta snapshots" in the background task, since
that could lead to data loss. For example, say we did "delta
snapshots" in the background, which removes the histogram samples from
the underlying storage. If the browser was shut down by the user and
the finalized log was never stored, then the samples would be lost.
Instead, in the background task, we take a snapshot of the samples
(without removing them from the underlying storage) and fill them into
the log. We only remove the samples from the underlying when we're
back on the main thread and stored the finalized log in Local State.

There is a feature to change this behaviour:
"MetricsServiceDeltaSnapshotInBg". When enabled, this performs full
"delta snapshots" in the background task (i.e. samples are removed
from the underlying storage and filled into the log). To make sure
there is no data loss, before the background task is posted, we create
a "ScopedKeepAlive", which forces the browser to stay alive (process
won't shut down). We destroy the ScopedKeepAlive after we're back on
the main thread and stored the finalized log in Local State.

However, according to crbug.com/1490916, there can be a short period
of time where we try to create a periodic ongoing log, but the browser
shutdown has already started. When this happens, creating a
ScopedKeepAlive triggers a CHECK. This CL makes it so that if shutdown
has already started, instead of creating a ScopedKeepAlive, do the
work synchronously instead.

This solution was discussed with in scheduler-dev:
https://groups.google.com/a/chromium.org/g/scheduler-dev/c/A1d7c-sGMmk/m/jWkM5Dh3AwAJ

And seems like a similar solution already exists in Sync:
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/background_sync/background_sync_delegate_impl.cc;l=63-68;drc=142b92dd3032947a049d244cded805617ea9b994


Bug: 1464694,1490916
Change-Id: If87bada9888574a79a25e99b60edd8121fc6d408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5059153
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229376}