Add segmentation_platform config parser fuzzer.
Noticed a gap for this parser in the fuzzing coverage dashboard which
could be trivially filled like this. I don't know whether the input here
is regarded as trusted or not, but we may as well fuzz it.

This fixes one failure that the fuzzer found immediately, which isn't
security-relevant.

Change-Id: I7ecf429ed4a6421bc16ade42e6d94defd73dbc34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5048732
Reviewed-by: Siddhartha S <ssid@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229392}