[RWS] Introduce SetsMutation type to uphold invariants
This does not change any behavior; it just introduces a new type that
serves as a "proof" of the invariants that
GlobalFirstPartySets::ComputeConfig relies on. This makes the
FirstPartySetParser more directly responsible for ensuring those
invariants hold.

(A side benefit of this is that if the parser ever fails to ensure
those invariants, now the stacktrace will implicate the parser
directly, since the parser is what now creates the SetsMutation. We
can take advantage of this by writing a new fuzzer to verify that the
parser ensures these properties; this will be done in a followup CL.)

Change-Id: Iea40068f98fe227abc84979f3b88b0b19f522a33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5059159
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Auto-Submit: Chris Fredrickson <cfredric@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229426}