[Tether][1/n] Thread NetworkHandler down to WifiHotspotConnector
This is part of a series of changes to introduce Wifi Connection
retries to WifiHotspotConnector.

This change threads NetworkHandler down to WifiHotspotConnector.
NetworkHandler will replace NetworkConnect as the mechanism for forming
a connection, as it offers a more robust interface for connections
(including fielding errors)

TEST=Existing Tests

Change-Id: I075f7d2d08bf7d1d9bebded0c15b0bc471071672
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5029219
Reviewed-by: Jack Shira <jackshira@google.com>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Joe Antonetti <joeantonetti@google.com>
Cr-Commit-Position: refs/heads/main@{#1229466}