chromeos: Add AppListSearchBrowserTest.SearchBuiltInApps
As part of go/crosier-conversion we're porting Tast tests that run in
the chromium CQ/CI to C++ GTest-based tests. This is a port of the test
launcher.SearchBuiltInApps. It doesn't have any platform dependencies,
so I ported it to browser_tests instead of chromeos_integration_tests.

Bug: b:312521316
Test: browser_tests --gtest_filter=AppListSearchBrowserTest.*
Change-Id: I5929b749da5cc07a9d2a5956e68707b03f58f48e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5038850
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229467}