Fix SitePerProcessBrowserTests affected by isolated, sandboxed iframes.
The SitePerProcessBrowserTest test suite contains three tests where the
use of sandboxed frames changes the output from DepictFrameTree due to
(1) the inclusion of port numbers in sandboxed origins, and (2) entries
from soon-to-be-deleted RenderFrameHosts still being around at the time
that DepictFrameTree is called. Also, due to the additional origins
caused by the sandboxed frames, labels in the DepictFrameTree output
(which are designed to remain unique across the entire test) can
change when IsolateSandboxedIframes is enabled.

This CL updates the affected tests to address all of these issue.
DepictFrameTree was modified to suppress port numbers in site
descriptions, expected output for per-site and per-origin modes was
added, and RenderFrameHostDeleted observers were added when
needed (this varies depending on which isolation mode, if any, is
in use). Per-document expected output will be added in a future
CL, and the tests will be parameterized to run all modes.

Bug: 510122
Change-Id: If6d6d7c9ad787e82eef4894042ff09eec5918be9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5013994
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: W. James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229490}