[shared storage] For SelectURL()/Run(), reject the promise when `document_service_` has gone
It was believed that `document_service_` cannot be null when
SharedStorageWorkletHost::SelectURL/Run is invoked, because the worklet
host mojom channel is associated with the document service mojom
channel. But this did occur in reality, and the root cause is still
unknown (likely due to a race). Therefore, we reject the promise
before doing the CHECK in the permission checks. In a follow-up,
we may add UMA to track how often it occurs.

Drive-by: Also remove the DumpWithoutCrashing() in
SharedStorageWorkletHost::Run() when `page_` is null, as it doesn't
provide helpful information, and previously we already removed the
DumpWithoutCrashing() for SelectURL().

Bug: 1502120, 1501559
Change-Id: Ie2d6f198193f1203fd9de893582fef482a846ecb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5063750
Reviewed-by: Cammie Smith Barnes <cammie@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229496}