overview: Rework how number of items are tracked
The old method doesn't account for snap groups or drop target, which
are technically overview items but have 2 and 0 windows respectively.
It was also needed to be updated each time we added or removed an item,
which is fragile.

Also, moves a for testing function into a test only file.

Test: existing tests
Change-Id: I28ccb3511d35c094986b5012e6114253df3d3f6e
Fixed: b/307598826
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5044515
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229515}