webview: shut down old cookie store after switch.
When switching from the temporary cookie store (used when CookieManager
is required before Chromium has been initialised) to the Mojo cookie
store that's used under normal operation, shut down the cookie store
once we've flushed it so that it closes the SQLite database properly,
and shut down the backend thread it was using to execute store jobs.

Also, lazily create the backend thread, as it's not needed in the case
where we don't create the temporary store in the first place.

The client thread is used by CookieManager in both implementations, so
its behaviour is not changed.

Fixed: 1062726
Change-Id: Ie1cc1afa22e91e4084b70f0aa8f1a178485cecc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095344
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Richard (Torne) Coles <torne@chromium.org>
Auto-Submit: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229539}