[sidepanelpinning] prevent button highlight from disappearing due to other interactions.
If using Button::AddAnchorHighlight on a toolbar button the highlight
would disappear if the context menu was opened then closed. To prevent
this this cl updates ToolbarButton to use Button::AddAnchorHighlight for
the context menu. The highlight was also disappearing after dragging due
to the ways Button::SetState gets called since the hover animation
affects the highlight state. To prevent this, we will check for the
anchor highlight count to make sure the highlight is correct after
state change.

Bug: b:313228038, 1428097
Change-Id: Iafdb4bd90c6d9b2bc439e93ec96aa40dd5fac30b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5058782
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229562}