Fix dangling pointer in cursor manager test
The test requires manual notification to the cursor manager when the
cursor is destroyed. This is typically handled by aura host view, which
the test host view does not inherit from. Other tests all explicitly
notify the cursor manager when the view is about to be destroyed.

Fixed: 1471085
Change-Id: Ifc1b6499faf98abf8d9330f7df7d5809f2f4805a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5011298
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1229579}