Check NativeThemeFluent doesn't have css colors when HC is on.
Painting functions in NativeThemeFluent depend on the correct input
from its callers to get the colors for the scrollbars' parts. If
CSS colors are enabled at the same time as forced colors, the function
would override the forced colors in favor of the css ones.

This CL adds CHECKS in the painting functions to make sure that this
situation doesn't occur.

Bug: 1495415
Change-Id: I42d520739d0cb822ce18e00b0b15ec57a4f306da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5038216
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1229642}