media/gpu: Add VP9SVCLayers that separate get and update function
The original VP9SVCLayers, VP9SVClayersStateful, updates the SVC state
during a function of getting a metadata for the frame. The new
VP9SVCLayers added in this CL doesn't update the state in getting
the metadata and instead have a function to update the state. This
enables to get the metadata multiple times for the same frame. It is
helpful to drop the input frame or re-encode the frame.
It is a client responsibility to handle the dynamic SVC configuration
change like L3T2 -> L2T3 and reset the SVC state for keyframe.

Bug: b:308046336
Test: media_unittests
Change-Id: If8e56e651e844b37a7d3299d864d8365e03363af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5040172
Reviewed-by: Justin Green <greenjustin@google.com>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229729}