Update DCHECKs in touchbar suggestion labels
The metadata project causes this to crash, because metadata was not
implemented in the autofill touch bar. So the value can end up being 2
due to metadata for both DCHECKs.

This CL updates the DCHECKs to be CHECKs as per Chromium style, and
ensures that they account for label updates from the metadata project as
well.

Fixed: 1503221
Change-Id: I083ad589f88b083b7e44152b90845f0d8bfdff3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5040900
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Alexander Tekle <alexandertekle@google.com>
Reviewed-by: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229764}