Fix Visual Studio visualizer for `BasicPersistent`
`BasicPersistent` uses a pointer, not `RawPointer`, stopped using `RawPointer`, and the `ptr_`
is no longer needed.  This patch changes the Visual Studio
visualizer to match the change.

Verified as per the README.md[1]. Note that this file wasn't
maintained for a while, and there are other warnings. This CL fixes
the warnings for `BasicPersistent`, but others are not covered.

[1] https://source.chromium.org/chromium/chromium/src/+/main:tools/win/DebugVisualizers/README.md

Change-Id: Idc6fbca18f27d433524d25a18c6a6ad2296ef247
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5058690
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229768}