Ensure CloseTaskHandler::DoDeferredClose does nothing if WebView is already closing
It's possible for CloseTaskHandler::DoDeferredClose to be called after
the WebView is in the middle of closing and Page::WillBeDestroyed is
already called, e.g. when the unload handler triggered from there
calls window.close(). In this case we shouldn't try to reach into the
WebViewImpl, and just do nothing.

This CL makes it so that CloseTaskHandler handles that case, and also
tidies up all the loose ends around setting/handling off
CloseTaskHandler to ensure that it's not created at awkward timings
as much as possible.

Bug: 1499519
Change-Id: Ie24758bc95d58eb93058c9850c7f399fb11b6384
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5058796
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229808}