Fix IsCreditCardUploadEnabled() wrt transport mode
IsCreditCardUploadEnabled() decide whether credit card save can be
offered, depending on whether address Autofill is enabled (and various
other things). This check depends on kSyncDecoupleAddressPaymentSettings
and kSyncEnableContactInfoDataTypeInTransportMode - feature flags
controlling whether the address and payments settings toggles are
coupled, and whether address sync is available in transport mode.

Currently, the logic is wrong on iOS, as outlined in the bug.

This CL simplifies the check to
bool IsCreditCardUploadEnabled() {
  ...
  bool syncing_or_addresses_in_transport_mode =
      sync_service->IsSyncFeatureActive() ||
      base::FeatureList::IsEnabled(
          kSyncEnableContactInfoDataTypeInTransportMode);
  if (syncing_or_addresses_in_transport_mode &&
      !sync_service->GetUserSettings()->GetSelectedTypes().Has(
          UserSelectableType::kAutofill)) {
    ...
    return false;
  }
}

This is correct in every scenario:
- kSyncEnableContactInfoDataTypeInTransportMode disabled
  - Upload possible for all users in transport mode.
  - In sync mode, upload possible for all users that haven't explicitly
    disabled "Addresses and more" in sync settings.
    When kSyncDecoupleAddressPaymentSettings is disabled, this is
    effectively all users, since they otherwise can't have payment sync
    enabled.
- kSyncEnableContactInfoDataTypeInTransportMode enabled (in particular,
  the current state on iOS)
  - No distinction between syncing and transport mode users.
  - Like in the case above, upload possible for all users that haven't
    explicitly disabled address sync. Whether this is possible at all
    depends on kSyncDecoupleAddressPaymentSettings.

----

Some comments about the current logic: Currently, the code checks for
  base::FeatureList::IsEnabled(kSyncDecoupleAddressPaymentSettings)
    ? sync_service->GetUserSettings()->GetSelectedTypes().Has(
        UserSelectableType::kAutofill)
    : sync_service->GetActiveDataTypes().Has(AUTOFILL_PROFILE)

That is, depending on kSyncDecoupleAddressPaymentSettings, either
default to the the status quo prior to crrev.com/c/4941899, or check for
the UserSelectableType.

In that sense, this CL regresses the behavior of checking for
AUTOFILL_PROFILE for syncing users. This is intended and arguably
desirable: AUTOFILL_WALLET_DATA can only be enabled when
AUTOFILL_PROFILE is enabled. This is already checked a few lines
above by the code. Prior to kSyncDecoupleAddressPaymentSettings, this
entire code block was only reachable for:
- Custom passphrase users, that haven't entered their passphrase yet.
  In this case, card upload is disabled due to a check further down in
  the function anyways.
- For users where AUTOFILL_PROFILE for some reason failed to start. This
  happens extremely rarely, because AUTOFILL_PROFILE doesn't have
  preconditions. Arguably, when AUTOFILL_PROFILE failed to start for
  infrastructure reasons, there's no need to disable upload, since the
  users explicitly has address syncing enabled from settings.

Fixed: 1504739
Change-Id: Id1b0ff4b1707a50da6ef229c43a2962268b16282
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5054023
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Olivia Saul <jsaul@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1229848}