[Code Health][iOS] Move ios/c/b/follow to model subfolder
To simplify the folder hierarchy on iOS and have a clearer layer
separation, all the files in ios/c/b/<foo>/ will be moved to
ios/c/b/<foo>/model/ to showcase that they are model files.

We use an hybrid approach for migrating upstream and downstream
smoothly. Here we keep dependencies (only a few) used by internal in
the  legacy folder, then we will change internal to use the new model/ folder for these dependencies, which weren't yet migrated. The
upstream code will use a forwarder header in model/ that forwards to
the remaining definitions in the legacy folder, up until the
migration is fully completed. Upstream clients of the ios/c/b/follow
modules are already fully migrated to ios/c/b/follow/model.

This CL is part of the chain of CLs for the model/ migration:
  #1 Migrate downstream and use forwarder where needed: THIS
  #2 Use model/ in internal: crrev.com/i/6706632 (tested)
  #3 Finish the migration of the not yet migrated dependencies.

Bug: 1319852
Change-Id: I3af1d13beb12476fc7e798f79532928894f3f3bb
Low-Coverage-Reason: Migration CL
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5059496
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: edchin <edchin@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1229877}