[//gpu] Move ownership of GpuMemoryBuffer into ClientSI
This CL changes ownership of the GpuMemoryBuffers associated with
MappableSharedImages from the SII implementations that create those
MappableSharedImages to the created ClientSharedImages. The lifetimes
in all cases are unchanged with one important assumption: that callers
of CreateSharedImage(BufferUsage) end up calling
DestroySharedImage(ClientSI) rather than DestroySharedImage(Mailbox)
[*]. This is true for all existing clients, and we will ensure that it
continues to be true as we convert more clients to MappableSharedImage.
(Note: the one exception is if the entire SII implementation gets torn
down before the client calls DestroySharedImage(ClientSI) on a given
ClientSI. We will need to handle that in the fullness of time, but in
the meantime, clients already can't effectively do anything with the
ClientSI/Mailbox once the SII is torn down).

*  This is important because the destruction of the GMB now occurs when
the ClientSI is destroyed. If the client calls
DestroySharedImage(ClientSI), then the GMB is guaranteed to be destroyed
when the Mailbox is destroyed as is happening now. If the client calls
DestroySharedImage(Mailbox), the destruction of the GMB will occur
whenever the client happens to drop the ClientSI. This would presumably
still be OK as the client needs to have the ClientSI to call
MapSharedImage() in any case, but it's harder to reason about.

Bug: 1500611
Change-Id: I1dcd0943bc840ee35db2e870e27d65cb9dd0695a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5029477
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229892}